Migrate radio-group from /docs/components/ to /components/
Move all examples (Basic, Form, Card) from the old documentation page
to the new reference page, delete the old page, and update all routes,
navigation links, and E2E tests to use /components/radio-group.
