--- site/ui/pages/radio-group.tsx
@@ -1,245 +1,0 @@
-/**
- * RadioGroup Documentation Page
- */
-
-import {
-  RadioGroupBasicDemo,
-  RadioGroupFormDemo,
-  RadioGroupCardDemo,
-} from '@/components/radio-group-demo'
-import {
-  DocPage,
-  PageHeader,
-  Section,
-  Example,
-  PropsTable,
-  PackageManagerTabs,
-  type PropDefinition,
-  type TocItem,
-} from '../components/shared/docs'
-import { getNavLinks } from '../components/shared/PageNavigation'
-
-// Table of contents items
-const tocItems: TocItem[] = [
-  { id: 'installation', title: 'Installation' },
-  { id: 'examples', title: 'Examples' },
-  { id: 'basic', title: 'Basic', branch: 'start' },
-  { id: 'form', title: 'Form', branch: 'child' },
-  { id: 'card', title: 'Card', branch: 'end' },
-  { id: 'api-reference', title: 'API Reference' },
-]
-
-// Code examples - Preview (Form Demo)
-const formCode = `"use client"
-
-import { createSignal, createMemo } from "@barefootjs/dom"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
-
-export function RadioGroupFormDemo() {
-  const [notifyType, setNotifyType] = createSignal("all")
-  const [theme, setTheme] = createSignal("system")
-
-  const summary = createMemo(() =>
-    \`Notifications: \${notifyType()}, Theme: \${theme()}\`
-  )
-
-  return (
-    <div className="space-y-6">
-      <div className="space-y-3">
-        <h4 className="text-sm font-medium leading-none">Notify me about...</h4>
-        <RadioGroup defaultValue="all" onValueChange={setNotifyType}>
-          <div className="flex items-center space-x-2">
-            <RadioGroupItem value="all" />
-            <span className="text-sm leading-none">All new messages</span>
-          </div>
-          <div className="flex items-center space-x-2">
-            <RadioGroupItem value="mentions" />
-            <span className="text-sm leading-none">Direct messages and mentions</span>
-          </div>
-          <div className="flex items-center space-x-2">
-            <RadioGroupItem value="none" />
-            <span className="text-sm leading-none">Nothing</span>
-          </div>
-        </RadioGroup>
-      </div>
-      <div className="space-y-3">
-        <h4 className="text-sm font-medium leading-none">Theme</h4>
-        <RadioGroup defaultValue="system" onValueChange={setTheme}>
-          <div className="flex items-center space-x-2">
-            <RadioGroupItem value="light" />
-            <span className="text-sm leading-none">Light</span>
-          </div>
-          <div className="flex items-center space-x-2">
-            <RadioGroupItem value="dark" />
-            <span className="text-sm leading-none">Dark</span>
-          </div>
-          <div className="flex items-center space-x-2">
-            <RadioGroupItem value="system" />
-            <span className="text-sm leading-none">System</span>
-          </div>
-        </RadioGroup>
-      </div>
-      <div className="text-sm text-muted-foreground pt-2 border-t">
-        {summary()}
-      </div>
-    </div>
-  )
-}`
-
-const basicCode = `"use client"
-
-import { createSignal } from "@barefootjs/dom"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
-
-export function RadioGroupBasicDemo() {
-  const [density, setDensity] = createSignal("default")
-
-  return (
-    <div className="space-y-4">
-      <RadioGroup defaultValue="default" onValueChange={setDensity}>
-        <div className="flex items-center space-x-2">
-          <RadioGroupItem value="default" />
-          <span className="text-sm font-medium leading-none">Default</span>
-        </div>
-        <div className="flex items-center space-x-2">
-          <RadioGroupItem value="comfortable" />
-          <span className="text-sm font-medium leading-none">Comfortable</span>
-        </div>
-        <div className="flex items-center space-x-2">
-          <RadioGroupItem value="compact" />
-          <span className="text-sm font-medium leading-none">Compact</span>
-        </div>
-      </RadioGroup>
-      <div className="text-sm text-muted-foreground pt-2 border-t">
-        Selected: {density()}
-      </div>
-    </div>
-  )
-}`
-
-const cardCode = `"use client"
-
-import { createSignal } from "@barefootjs/dom"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
-
-const plans = [
-  { value: "startup", name: "Startup", price: "$29", description: "For small teams getting started" },
-  { value: "business", name: "Business", price: "$99", description: "For growing companies" },
-  { value: "enterprise", name: "Enterprise", price: "$299", description: "For large organizations" },
-]
-
-export function RadioGroupCardDemo() {
-  const [plan, setPlan] = createSignal("startup")
-
-  return (
-    <div className="space-y-4">
-      <RadioGroup defaultValue="startup" onValueChange={setPlan} class="grid-cols-1 sm:grid-cols-3">
-        {plans.map((p) => (
-          <div key={p.value} className="relative">
-            <div className="flex items-start space-x-3 rounded-lg border p-4 hover:bg-accent/50 cursor-pointer">
-              <RadioGroupItem value={p.value} />
-              <div className="space-y-1">
-                <span className="text-sm font-medium leading-none">{p.name}</span>
-                <p className="text-xl font-bold text-foreground">
-                  {p.price}<span className="text-sm font-normal text-muted-foreground">/mo</span>
-                </p>
-                <p className="text-sm text-muted-foreground">{p.description}</p>
-              </div>
-            </div>
-          </div>
-        ))}
-      </RadioGroup>
-      <div className="text-sm text-muted-foreground pt-2 border-t">
-        Selected plan: {plan()}
-      </div>
-    </div>
-  )
-}`
-
-// Props definitions
-const radioGroupProps: PropDefinition[] = [
-  {
-    name: 'defaultValue',
-    type: 'string',
-    description: 'The initial selected value for uncontrolled mode.',
-  },
-  {
-    name: 'value',
-    type: 'string',
-    description: 'The controlled selected value. When provided, the component is in controlled mode.',
-  },
-  {
-    name: 'onValueChange',
-    type: '(value: string) => void',
-    description: 'Event handler called when the selected value changes.',
-  },
-  {
-    name: 'disabled',
-    type: 'boolean',
-    defaultValue: 'false',
-    description: 'Whether the entire radio group is disabled.',
-  },
-]
-
-const radioGroupItemProps: PropDefinition[] = [
-  {
-    name: 'value',
-    type: 'string',
-    description: 'The value of this radio item. Required.',
-  },
-  {
-    name: 'disabled',
-    type: 'boolean',
-    defaultValue: 'false',
-    description: 'Whether this radio item is disabled.',
-  },
-]
-
-export function RadioGroupPage() {
-  return (
-    <DocPage slug="radio-group" toc={tocItems}>
-      <div className="space-y-12">
-        <PageHeader
-          title="Radio Group"
-          description="A set of checkable buttons where only one can be checked at a time."
-          {...getNavLinks('radio-group')}
-        />
-
-        {/* Preview */}
-        <Example title="" code={formCode}>
-          <RadioGroupFormDemo />
-        </Example>
-
-        {/* Installation */}
-        <Section id="installation" title="Installation">
-          <PackageManagerTabs command="barefoot add radio-group" />
-        </Section>
-
-        {/* Examples */}
-        <Section id="examples" title="Examples">
-          <div className="space-y-8">
-            <Example title="Basic" code={basicCode}>
-              <RadioGroupBasicDemo />
-            </Example>
-
-            <Example title="Form" code={formCode}>
-              <RadioGroupFormDemo />
-            </Example>
-
-            <Example title="Card" code={cardCode}>
-              <RadioGroupCardDemo />
-            </Example>
-          </div>
-        </Section>
-
-        {/* API Reference */}
-        <Section id="api-reference" title="API Reference">
-          <h3 className="text-lg font-semibold mb-4">RadioGroup</h3>
-          <PropsTable props={radioGroupProps} />
-          <h3 className="text-lg font-semibold mb-4 mt-8">RadioGroupItem</h3>
-          <PropsTable props={radioGroupItemProps} />
-        </Section>
-      </div>
-    </DocPage>
-  )
-}
