+++ site/ui/renderer.tsx
@@ -103,7 +103,7 @@ const menuEntries: SidebarEntry[] = [
       { title: 'Pagination', href: '/docs/components/pagination' },
       { title: 'Popover', href: '/docs/components/popover' },
       { title: 'Progress', href: '/docs/components/progress' },
-      { title: 'Radio Group', href: '/docs/components/radio-group' },
+      { title: 'Radio Group', href: '/components/radio-group' },
       { title: 'Resizable', href: '/docs/components/resizable' },
       { title: 'Scroll Area', href: '/docs/components/scroll-area' },
       { title: 'Select', href: '/components/select' },
